Revert "Separate detailed styles for entry list"
This reverts commit df226bcb5f4db342ee2f3979d6426931f0d9243d.
