+++ app/components/concerns/EntryList/index.ts
@@ -1,1 +1,2 @@
+export { EntryArticle } from "./EntryArticle";
 export { EntryList } from "./EntryList";
